Revert "#141409277 invite friends #141409277 effective onbaording"
