+++ app/views/includes/foot.jade
@@ -3,15 +3,14 @@ script.
       (function(d,t){var g=d.createElement(t),s=d.getElementsByTagName(t)[0];
       g.src='//www.google-analytics.com/ga.js';
       s.parentNode.insertBefore(g,s)}(document,'script'));
+script(type='text/javascript', src='/lib/jquery/jquery.js')
 script(type='text/javascript', src='/lib/jquery/jquery.min.js')
-script(type='text/javascript', src='/js/notify.js')
 script(type='text/javascript', src='/lib/ResponsiveSlides/responsiveslides.min.js')
 script(type='text/javascript', src='/lib/underscore/underscore-min.js')
 
-
-
 //	Bootstrap
 script(type='text/javascript', src='/lib/bootstrap/dist/js/bootstrap.js')
+
 //AngularJS
 script(type='text/javascript', src='https://code.angularjs.org/1.1.5/angular.js')
 script(type='text/javascript', src="https://rawgithub.com/gsklee/ngStorage/master/ngStorage.js")
@@ -21,7 +20,6 @@ script(type='text/javascript', src='https://code.angularjs.org/1.1.5/angular-coo
 //	Angular UI
 script(type='text/javascript', src='/lib/angular-bootstrap/ui-bootstrap-tpls.js')
 script(type='text/javascript', src='/lib/angular-ui-utils/modules/route/route.js')
-script(type='text/javascript', src='/lib/angular-intro.js/build/angular-intro.min.js')
 
 //	Application Init
 script(type='text/javascript', src='/js/init.js')
@@ -39,11 +37,6 @@ script(type='text/javascript', src='/js/controllers/index.js')
 script(type='text/javascript', src='/js/controllers/header.js')
 script(type='text/javascript', src='/js/controllers/game.js')
 script(type='text/javascript', src='/js/init.js')
-
-//IntroJS
-script(type='text/javascript', src='/lib/intro.js/intro.js')
-//- script(type='text/javascript', src='/js/intro.js')
-script(type='text/javascript', src='/js/game_tour.js')
 
 //	Socket.io Client Library
 script(type='text/javascript', src='/socket.io/socket.io.js')
