+++ app/views/includes/head.jade
@@ -27,13 +27,11 @@ head
   meta(property='fb:admins', content='APP_ADMIN')
 
   link(rel='stylesheet', href='/lib/bootstrap/dist/css/bootstrap.css')
-  link(rel='stylesheet', href='/lib/font-awesome/css/font-awesome.min.css')
-
+  link(href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css", rel="stylesheet")
   link(rel='stylesheet', href='/css/common.css')
   link(rel='stylesheet', href='/css/animate.css')
   link(rel='stylesheet', href='/css/style.css')
   link(rel='stylesheet', href='/css/homestyle.css')
-  link(rel='stylesheet', href='/lib/intro.js/minified/introjs.min.css') 
   link(rel='stylesheet', href='/css/game-play.css')
