+++ bower.json
@@ -12,7 +12,6 @@
   "exportsOverride": {
     "bootstrap": {
       "css": "dist/css/*",
-      
       "js": "dist/js/*",
       "fonts": "dist/fonts/*"
     },
