+++ public/css/style.css
@@ -17,7 +17,7 @@
 }
 
  .fa-facebook {
-   color: #fff;
+   color: #3b5998;
  }
 
  .fa-github {
@@ -1364,8 +1364,8 @@
  }
 
  #searchbox {
-  width: 200px;
-  height: 50px;
+  width: 400px;
+  height: 45px;
   vertical-align: middle;
   white-space: nowrap;
   position: relative;
