+++ public/js/app.js
@@ -1,4 +1,4 @@
-angular.module('mean', ['ngCookies', 'ngResource', 'ui.bootstrap', 'ui.route', 'mean.system', 'mean.directives', 'angular-intro'])
+angular.module('mean', ['ngCookies', 'ngResource', 'ui.bootstrap', 'ui.route', 'mean.system', 'mean.directives'])
   .config(['$routeProvider',
       function($routeProvider) {
           $routeProvider.
@@ -25,9 +25,6 @@ angular.module('mean', ['ngCookies', 'ngResource', 'ui.bootstrap', 'ui.route', '
           }).
           when('/choose-avatar', {
             templateUrl: '/views/choose-avatar.html'
-          }).
-          when('/gametour', {
-            templateUrl: '/views/tour.html'
           }).
           otherwise({
             redirectTo: '/'
