+++ public/js/controllers/index.js
@@ -1,6 +1,6 @@
 
 angular.module('mean.system')
-  .controller('IndexController',
+.controller('IndexController',
   [
     '$scope',
     'Global',
@@ -36,76 +36,46 @@ angular.module('mean.system')
       $scope.signUp = () => {
         $scope.showMessage = '';
         $http.post('/api/auth/signup', JSON.stringify($scope.formData))
-          .success((data) => {
-            if (data.success === true) {
-              $window.localStorage.setItem('user-token', data.token);
-              $window.location.href = '/gametour';
-            } else {
-              $scope.showMessage = data.message;
-              angular.element(document.querySelector('#name')).val('');
-              angular.element(document.querySelector('#email')).val('');
-              angular.element(document.querySelector('#password')).val('');
-            }
-          }).error((error, status) => {
-            $scope.showMessage = `${status} : ${error}`;
-            angular.element(document.querySelector('#name')).val('');
-            angular.element(document.querySelector('#email')).val('');
-            angular.element(document.querySelector('#password')).val('');
-          });
+      .success((data) => {
+        if (data.success === true) {
+          $window.localStorage.setItem('user-token', data.token);
+          $window.location.href = '/';
+        } else {
+          $scope.showMessage = data.message;
+          angular.element(document.querySelector('#name')).val('');
+          angular.element(document.querySelector('#email')).val('');
+          angular.element(document.querySelector('#password')).val('');
+        }
+      }).error((error, status) => {
+        $scope.showMessage = `${status} : ${error}`;
+        angular.element(document.querySelector('#name')).val('');
+        angular.element(document.querySelector('#email')).val('');
+        angular.element(document.querySelector('#password')).val('');
+      });
       };
 
       $scope.signIn = () => {
         $scope.showMessage = '';
         $http.post('/api/auth/signin', JSON.stringify($scope.formData))
-          .success((data) => {
-            if (data.success === true) {
-              $window.localStorage.setItem('user-token', data.token);
-              $window.location.href = '/';
-            } else {
-              $scope.showMessage = data.message;
-              angular.element(document.querySelector('#email')).val('');
-              angular.element(document.querySelector('#password')).val('');
-            }
-          }).error((error, status) => {
-            $scope.showMessage = `${status} : ${error}`;
+        .success((data) => {
+          if (data.success === true) {
+            $window.localStorage.setItem('user-token', data.token);
+            $window.location.href = '/';
+          } else {
+            $scope.showMessage = data.message;
             angular.element(document.querySelector('#email')).val('');
             angular.element(document.querySelector('#password')).val('');
-          });
+          }
+        }).error((error, status) => {
+          $scope.showMessage = `${status} : ${error}`;
+          angular.element(document.querySelector('#email')).val('');
+          angular.element(document.querySelector('#password')).val('');
+        });
       };
-      if (window.user !== null) {
-        if (window.user._id !== null) {
-          $scope.LoadNotifications = () => {
-            const userId = window.user._id;
-            $http.post('/api/notify', { user_id: userId })
-              .success((response) => {
-                $scope.allNotification = response;
-                $scope.count = $scope.allNotification.length;
-                console.log(response);
-              }, (error) => {
-                console.log(error);
-              }
-              );
-          };
 
-          $scope.readNotifications = (notifyId) => {
-            const userId = window.user._id;
-            $http.post('/api/read', { user_id: userId, notifyId })
-              .success((response) => {
-                if (response.succ) {
-                  $scope.LoadNotifications();
-                }
-              }, (error) => {
-                console.log(error);
-              }
-              );
-          };
-          $scope.LoadNotifications();
-        }
-      }
-      // console.log(window);
       $scope.avatars = [];
       AvatarService.getAvatars()
-        .then((data) => {
-          $scope.avatars = data;
-        });
+      .then((data) => {
+        $scope.avatars = data;
+      });
     }]);
