--- public/js/notify.js
@@ -1,21 +1,0 @@
-$(document).ready(function()
-{
-$("#notificationLink").click(function()
-{
-$("#notificationContainer").fadeToggle(300);
-$("#notification_count").fadeOut("slow");
-return false;
-});
-
-//Document Click
-$(document).click(function()
-{
-$("#notificationContainer").hide();
-});
-//Popup Click
-$("#notificationContainer").click(function()
-{
-return false
-});
-
-});
