+++ public/views/app.html
@@ -1,24 +1,24 @@
 <div id="gameBegun" class="modal fade" role="dialog">
-  <div class="modal-dialog">
-    <!-- Modal content-->
-    <div class="modal-content">
-      <div class="modal-header">
-        <h4 class="modal-title" style="color: #800000">The game has begun, slowpoke. </h4>
-      </div>
-      <div class="modal-body">
-        <p>Go ahead and lick your wounds.</p>
-      </div>
-      <div class="modal-footer">
-        <button type="button" class="btn btn-default" data-dismiss="modal" onclick="window.location = '/'">Close</button>
+      <div class="modal-dialog">
+        <!-- Modal content-->
+        <div class="modal-content">
+          <div class="modal-header">
+            <h4 class="modal-title" style="color: #800000">The game has begun, slowpoke. </h4>
+          </div>
+          <div class="modal-body">
+            <p>Go ahead and lick your wounds.</p>
+          </div>
+          <div class="modal-footer">
+            <button type="button" class="btn btn-default" data-dismiss="modal" onclick="window.location = '/'">Close</button>
+          </div>
+        </div>
+
       </div>
 </div>
 
-
-  </div>
-</div>
 
 <div id="app-container" ng-controller="GameController">
-<div class="app-container-bg">
+<div class="slider">
   <div id="menu-container" class="nav-wrapper">
       <nav class="navbar" role="navigation">
         <a href="#!/">
@@ -27,11 +27,12 @@
             <span id="second-word"> for</span>
             <span id="third-word">Humanity</span>
           </div>
-          </a>
+        </a>
+
         <div class=" pull-right nav-wrapper">
         <a href="#" data-activates="mobile" class="button-collapse"><i class="material-icons"></i></a>
               <ul class="navbar-nav">
-                <li ng-hide="showOptions" class="dropdown" id="dropdown">
+                <li class="dropdown" id="dropdown">
                   <a href="#" class="dropdown-toggle" data-toggle="dropdown"><i class="glyphicon glyphicon-bell"></i>
                     <span ng-show="allNotification.length" class="label label-danger">{{ count }}</span>
                   </a>
@@ -48,6 +49,7 @@
                 </li>
                 <li id="tweet-container">
                   <a href="https://twitter.com/share" class="twitter-share-button" data-url="http://cfh.io" data-text="Cards for Humanity: A Game for Horrible People Desperately Trying to do Good" data-related="CFH_App"><img src="../images/tweet.png" alt="share-on-twitter" width="30px" height="30px"></a>
+                  <script>!function(d,s,id){var js,fjs=d.getElementsByTagName(s)[0],p=/^http:/.test(d.location)?'http':'https';if(!d.getElementById(id)){js=d.createElement(s);js.id=id;js.src=p+'://platform.twitter.com/widgets.js';fjs.parentNode.insertBefore(js,fjs);}}(document, 'script', 'twitter-wjs');
                   </script>
                 </li>
               </ul>
